% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_auxiliaries.R
\name{agg_survey}
\alias{agg_survey}
\title{Aggregate surveys}
\usage{
agg_survey(survey, bycats, estimate_se = FALSE)
}
\arguments{
\item{survey}{Survey data file.}

\item{bycats}{Category for aggregation.}

\item{estimate_se}{Boolean indicating whether to estimate standard errors.}
}
\description{
Aggregate surveys
}
