% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ajustar_casillas_especiales.R
\name{ajustar_casillas_especiales}
\alias{ajustar_casillas_especiales}
\title{Ajusta el marco muestral y la muestra de acuerdo al criterio que se utilice}
\usage{
ajustar_casillas_especiales(muestra, marco_muestral, criterio)
}
\arguments{
\item{muestra}{La muestra con la que se realiza las estimaciones. Debe contener la columna LISTA_NOMINAL.}

\item{marco_muestral}{El marco muestral del que se extrae la muestra. Debe contener la columnas ID_ESTADO, ID_DISTRITO, SECCION, TIPO_CASILLA, EXT_CONTIGUA, ID_CASILLA, y LISTA_NOMINA. Las casillas especiales deben tener la letra 'S'.}

\item{criterio}{Criterio para ajustar la lista nominal de las casillas especiales, puesto que por su naturaleza no cuentan con lista nominal. Las opciones disponibles son '2018'.
\emph{2018}: Agregar la lista nominal máxima a las casillas especiales (750). Posteriormente, restar a todas las casillas de manera uniforme el incremento en la lista nominal.}
}
\value{

}
\description{
Ajusta el marco muestral y la muestra de acuerdo al criterio que se utilice
}
