% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99_deprecated_testing.R
\name{alt1_smpl}
\alias{alt1_smpl}
\title{Simple model test (Deprecated).}
\usage{
alt1_smpl(df)
}
\arguments{
\item{df}{Input data frame.}
}
\description{
Executes a simple linear model in the probit space.
}
