% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_testing.R
\name{alt1_smpl2}
\alias{alt1_smpl2}
\title{Observation level simple model test.}
\usage{
alt1_smpl2(input, tests)
}
\arguments{
\item{input}{Input set data frame.}

\item{tests}{Test set data frame.}
}
\value{
Returns a summary table with the model performance.
}
\description{
\code{alt1_smpl2} uses the results of a simple linear model in the probit
space estimated using the input set to estimate the observations described
by the test set. The differences are summarised by the mean squared error,
mean absolute deviation, and median absolute deviation.
}
\seealso{
Other testing functions: 
\code{\link{alt1_smpl3}()},
\code{\link{alt2_flat2}()},
\code{\link{alt2_flat3}()},
\code{\link{alt3_ltst3}()},
\code{\link{remove_latest}()},
\code{\link{remove_recent}()},
\code{\link{test_abc2}()},
\code{\link{test_abc}()}
}
\concept{testing functions}
