% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99_deprecated_testing.R
\name{alt2_flat}
\alias{alt2_flat}
\title{Flat model test (Deprecated).}
\usage{
alt2_flat(df)
}
\arguments{
\item{df}{Input data frame.}
}
\description{
Executes an intercept only model in the probit space.
}
