% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99_deprecated_testing.R
\name{alt3_ltst}
\alias{alt3_ltst}
\title{Latest model test (Deprecated).}
\usage{
alt3_ltst(df)
}
\arguments{
\item{df}{Input data frame.}
}
\description{
Executes a latest survey model in the probit space.
}
