% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_testing.R
\name{alt3_ltst3}
\alias{alt3_ltst3}
\title{Survey level latest model test.}
\usage{
alt3_ltst3(input, tests)
}
\arguments{
\item{input}{Input set data frame.}

\item{tests}{Test set data frame.}
}
\value{
Returns a summary table with the model performance.
}
\description{
\code{alt3_ltst3} uses the results of a "latest available value" model
estimated using the input set to estimate the completion rate indicator
of the observation year of each survey in the test set. The differences
are summarised by the mean squared error,mean absolute deviation, and
median absolute deviation.
}
\seealso{
Other testing functions: 
\code{\link{alt1_smpl2}()},
\code{\link{alt1_smpl3}()},
\code{\link{alt2_flat2}()},
\code{\link{alt2_flat3}()},
\code{\link{remove_latest}()},
\code{\link{remove_recent}()},
\code{\link{test_abc2}()},
\code{\link{test_abc}()}
}
\concept{testing functions}
