% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_plotting.R
\name{bkprj_plot}
\alias{bkprj_plot}
\title{Plot input backprojection (Deprecated)}
\usage{
bkprj_plot(df, cs, lvl = "prim", sx = "total")
}
\arguments{
\item{df}{Input data frame.}

\item{cs}{Countries to plot.}

\item{lvl}{Level to plot.}

\item{sx}{Sex to plot.}
}
\description{
Plot input backprojection (Deprecated)
}
