% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99_deprecated_testing.R
\name{cnsr_ltst}
\alias{cnsr_ltst}
\title{Censor latest (Deprecated).}
\usage{
cnsr_ltst(df)
}
\arguments{
\item{df}{Input data frame.}
}
\value{
Returns df with each test set observation having NA as its value.
}
\description{
This function sets the value of observations attributed to the most recent
survey for each country to NA. These censored values are used as a test set
with which the ABC model can be evaluated.
}
