% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_functions.R
\name{combine_regs}
\alias{combine_regs}
\title{Complete regional aggregation}
\usage{
combine_regs(df)
}
\arguments{
\item{df}{Input data frame.}
}
\value{
Data frame summarising aggregations.
}
\description{
\code{combine_regs} aggregates completion rate estimates based on SDG region,
income group, and SDG region x income group using population weights. The
results of each type of aggregation are then binded into a single data frame.
}
