% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compulsar.R
\name{compulsar}
\alias{compulsar}
\title{Title}
\usage{
compulsar(
  info,
  estado,
  raiz,
  carpeta_interna,
  carpeta_remesas,
  carpeta_salida,
  estimaciones_compulsadas,
  tiempo_max,
  intevalo
)
}
\arguments{
\item{carpeta_salida}{}
}
\value{

}
\description{
Title
}
