% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/en_vivo.R
\name{correr_conteo_rapido}
\alias{correr_conteo_rapido}
\title{Title}
\usage{
correr_conteo_rapido(
  carpeta_remesas,
  carpeta_INE,
  carpeta_interna,
  info_eleccion,
  candidatos,
  otros,
  marco_muestral,
  id_casilla,
  id_estrato,
  lista_nominal,
  n_sim = 10000,
  tipo_casilla = TIPO_CASILLA,
  resultados = NULL
)
}
\arguments{
\item{resultados}{}
}
\value{

}
\description{
Title
}
