% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99_deprecated_modeling.R
\name{country_rep}
\alias{country_rep}
\title{Country repeater (Deprecated).}
\usage{
country_rep(clist, lower, upper)
}
\arguments{
\item{clist}{Vector of countries.}

\item{lower}{Lower bound country.}

\item{upper}{Upper bound country.}
}
\description{
Function to prepare country list for slicing.
}
