% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crear_estratos.R
\name{crear_estratos1}
\alias{crear_estratos1}
\title{Title}
\usage{
crear_estratos1(bd, nombre, condicion, m_min, prefijo = "estratificacion")
}
\arguments{
\item{m_min}{}
}
\value{

}
\description{
Title
}
