% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluar_estratificacion.R
\name{evaluar_estratificacion}
\alias{evaluar_estratificacion}
\title{Title}
\usage{
evaluar_estratificacion(
  marco_muestral,
  n_muestras = 100,
  id_estratos,
  candidatos,
  fracc_muestreo = 0.1
)
}
\arguments{
\item{fracc_muestreo}{}
}
\value{

}
\description{
Title
}
