% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluar_estratificacion.R
\name{evaluar_estratificacion_censurada}
\alias{evaluar_estratificacion_censurada}
\title{Title}
\usage{
evaluar_estratificacion_censurada(
  marco_muestral,
  n_muestras = 100,
  id_estratos,
  candidatos,
  fracc_muestreo = 0.05,
  reg_muestras = F,
  modelo,
  guardar = F
)
}
\arguments{
\item{fracc_muestreo}{}
}
\value{

}
\description{
Title
}
