% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluar_estratificacion.R
\name{evaluar_estratificacion_precision}
\alias{evaluar_estratificacion_precision}
\title{Title}
\usage{
evaluar_estratificacion_precision(estimaciones, precision = 0.02)
}
\value{

}
\description{
Title
}
