% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99_deprecated_modeling.R
\name{fn_agg}
\alias{fn_agg}
\title{Aggregate function (Deprecated).}
\usage{
fn_agg(df, as, ...)
}
\arguments{
\item{df}{Input data frame.}

\item{as}{Ages?}

\item{...}{Pass through parameters.}
}
\description{
Aggregate function (Deprecated).
}
