% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_extraction.R
\name{get_bias}
\alias{get_bias}
\title{Extract Survey Bias}
\usage{
get_bias(df, raw_mcmc, rounds)
}
\arguments{
\item{df}{Input data frame.}

\item{raw_mcmc}{ABC model output.}

\item{rounds}{Survey rounds data frame.}
}
\value{
Data frame with columns for survey, iteration, variable, value, country,
  year, round, level, and sex.
}
\description{
Extracts survey bias iterations from the model output and attaches information
about the survey. Note that no summary statistics are computed.
}
\seealso{
Other extraction functions: 
\code{\link{get_drift}()},
\code{\link{get_muzero}()},
\code{\link{get_pars}()},
\code{\link{get_rhats}()},
\code{\link{get_yrepl}()}
}
\concept{extraction functions}
