% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_extraction.R
\name{get_drift}
\alias{get_drift}
\title{Extract long-term drift}
\usage{
get_drift(df, raw_mcmc)
}
\arguments{
\item{df}{Input data frame.}

\item{raw_mcmc}{ABC model output.}
}
\value{
Data frame with columns for country, value, and lower/upper bounds. The
  output does not specify the level and sex combination as it is assumed to be
  known by the user based on the inputs.
}
\description{
Extracts drift iterations from the model output and summarises the iterations
in a 90\% interval.
}
\seealso{
Other extraction functions: 
\code{\link{get_bias}()},
\code{\link{get_muzero}()},
\code{\link{get_pars}()},
\code{\link{get_rhats}()},
\code{\link{get_yrepl}()}
}
\concept{extraction functions}
