% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_functions.R
\name{get_muzero}
\alias{get_muzero}
\title{Extract Latent Space Intercept}
\usage{
get_muzero(df, raw_mcmc)
}
\arguments{
\item{df}{Input data frame.}

\item{raw_mcmc}{ABC model output.}
}
\value{
Data frame with columns for country, value, and lower/upper bounds. The
  output does not specify the level and sex combination as it is assumed to be
  known by the user based on the inputs.
}
\description{
Extracts the intercept term from the model output and summarises the iterations
in a 90\% interval. Note that the intercept for the ABC dataset corresponds to
1980.
}
\seealso{
Other extraction functions: 
\code{\link{get_bias}()},
\code{\link{get_drift}()},
\code{\link{get_pars}()},
\code{\link{get_rhats}()},
\code{\link{get_yrepl}()}
}
\concept{extraction functions}
