% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_extraction.R
\name{get_nsvar}
\alias{get_nsvar}
\title{Extract Non-Sampling Variance Samples.}
\usage{
get_nsvar(df, raw_mcmc)
}
\arguments{
\item{df}{Input data frame}

\item{raw_mcmc}{Stan model output}
}
\description{
Extracts samples for non-sampling variance and subsequently merges them by country
for input into the leave-survey-out generated quantities model.
}
\seealso{
Other extraction functions: 
\code{\link{get_bias}()},
\code{\link{get_drift}()},
\code{\link{get_jt_drift}()},
\code{\link{get_muzero}()},
\code{\link{get_parsamps}()},
\code{\link{get_pars}()},
\code{\link{get_rhats}()},
\code{\link{get_yrepl}()}
}
\concept{extraction functions}
