% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_extraction.R
\name{get_pars}
\alias{get_pars}
\title{Extract Model Parameters}
\usage{
get_pars(df, raw_mcmc)
}
\arguments{
\item{df}{Input data frame.}

\item{raw_mcmc}{ABC model output.}
}
\value{
Data frame with columns for survey, parameter, country, value, lower, upper,
  and level.
}
\description{
Extracts survey, late, very late, and age misreporting bias terms from the
model output, then presents the median and 90 percent intervals for each parameter.
}
\seealso{
Other extraction functions: 
\code{\link{get_bias}()},
\code{\link{get_drift}()},
\code{\link{get_jt_drift}()},
\code{\link{get_muzero}()},
\code{\link{get_nsvar}()},
\code{\link{get_parsamps}()},
\code{\link{get_rhats}()},
\code{\link{get_yrepl}()}
}
\concept{extraction functions}
