% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_extraction.R
\name{get_parsamps}
\alias{get_parsamps}
\title{Extract Parameter Samples.}
\usage{
get_parsamps(raw_mcmc, par)
}
\arguments{
\item{raw_mcmc}{Stan model output}

\item{par}{Parameter name}
}
\description{
Extracts samples for a specified parameter family to flow into the generated quantities
model.
}
\seealso{
Other extraction functions: 
\code{\link{get_bias}()},
\code{\link{get_drift}()},
\code{\link{get_jt_drift}()},
\code{\link{get_muzero}()},
\code{\link{get_nsvar}()},
\code{\link{get_pars}()},
\code{\link{get_rhats}()},
\code{\link{get_yrepl}()}
}
\concept{extraction functions}
