% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_extraction.R
\name{get_rhats}
\alias{get_rhats}
\title{Extract Rhats}
\usage{
get_rhats(df, level, sex)
}
\arguments{
\item{df}{Input data frame.}

\item{level}{Corresponding level.}

\item{sex}{Corresponding sex.}
}
\description{
Extracts Rhats for all parameters (base and transformed). Note that generated quantities
are excluded. The 9 parameters with the worst Rhats are returned.
}
\seealso{
Other extraction functions: 
\code{\link{get_bias}()},
\code{\link{get_drift}()},
\code{\link{get_jt_drift}()},
\code{\link{get_muzero}()},
\code{\link{get_nsvar}()},
\code{\link{get_parsamps}()},
\code{\link{get_pars}()},
\code{\link{get_yrepl}()}
}
\concept{extraction functions}
