% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_extraction.R
\name{get_yrepl}
\alias{get_yrepl}
\title{Extract observation replications.}
\usage{
get_yrepl(input, mcmc)
}
\arguments{
\item{input}{Input data frame.}

\item{mcmc}{ABC model output.}
}
\value{
Data frame containing the replications generated at each iteration
of the ABC model.
}
\description{
Extract observation replications.
}
\seealso{
Other extraction functions: 
\code{\link{get_bias}()},
\code{\link{get_drift}()},
\code{\link{get_muzero}()},
\code{\link{get_pars}()},
\code{\link{get_rhats}()}
}
\concept{extraction functions}
