% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graficar_en_vivo.R
\name{graficar_distribucion_ev}
\alias{graficar_distribucion_ev}
\title{Title}
\usage{
graficar_distribucion_ev(resultados, candidato, por_estrato = F)
}
\arguments{
\item{resultados}{}
}
\value{

}
\description{
Title
}
