% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graficar_en_vivo.R
\name{graficar_estimaciones_ev}
\alias{graficar_estimaciones_ev}
\title{Title}
\usage{
graficar_estimaciones_ev(
  resultados,
  candidatos,
  estimacion = "puntual",
  participacion = F
)
}
\arguments{
\item{candidatos}{}
}
\value{

}
\description{
Title
}
