% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_modelling.R
\name{init_drift}
\alias{init_drift}
\title{Initialize the drift parameter.}
\usage{
init_drift(ncountry, nyears, joint = FALSE)
}
\arguments{
\item{ncountry}{Number of countries.}
}
\value{
Initial values for the drift parameter
}
\description{
Due to the EMG prior on the drift parameter, the standard Unif(-2,2)
can produce values with effective probability 0 due to machine
precision limitations. Thus, we consider Unif(-0.5,2) instead.
}
