% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_contents.R
\name{isolate_tune}
\alias{isolate_tune}
\title{Isolate a tune, starting at a specific line.}
\usage{
isolate_tune(.lines, .start)
}
\arguments{
\item{.start}{}
}
\value{

}
\description{
Isolate a tune, starting at a specific line.
}
