% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_auxiliaries.R
\name{level_names}
\alias{level_names}
\title{Substitute levels with full titles.}
\usage{
level_names(...)
}
\arguments{
\item{...}{Pass through parameters.}
}
\description{
Substitute levels with full titles.
}
