% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_processing.R
\name{loo_influential}
\alias{loo_influential}
\title{Determine influential points.}
\usage{
loo_influential(df, loo1, threshold = 0.7)
}
\arguments{
\item{df}{Input data frame.}

\item{loo1}{LOO object.}

\item{threshold}{Pareto-K influential threshold}
}
\value{
The set of observations determined to be highly influential to parameter
estimation.
}
\description{
\code{loo_influential} uses the estimated Pareto-k values from the PSIS-LOO
method to determine which points are influential to model estimation. Note that
due to the model structure, some observations will by design produce high Pareto-k
values such as a3 points subject to late completion. While PSIS-LOO may not generate
a usable LOO estimate, it does point towards interesting observations.
}
