% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_processing.R
\name{loo_output}
\alias{loo_output}
\title{ABC model PSIS-LOO.}
\usage{
loo_output(mcmc)
}
\arguments{
\item{mcmc}{Raw MCMC output.}
}
\value{
LOO object.
}
\description{
Performs PSIS-LOO on the ABC model output.
}
