% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_functions.R
\name{loo_pit}
\alias{loo_pit}
\title{Plot ABC LOO-PIT}
\usage{
loo_pit(df, raw_mcmc, loo1)
}
\arguments{
\item{df}{Input data frame.}

\item{raw_mcmc}{ABC model output.}

\item{loo1}{LOO object.}
}
\description{
Plots the LOO-PIT for the ABC model output
}
