% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_processing.R
\name{loo_qq}
\alias{loo_qq}
\title{Plot ABC LOO-PIT}
\usage{
loo_qq(df, raw_mcmc, loo1)
}
\arguments{
\item{df}{Input data frame.}

\item{raw_mcmc}{ABC model output.}

\item{loo1}{LOO object.}
}
\description{
Plots the LOO-PIT QQ for the ABC model output
}
