% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_modelling.R
\name{mdl_compile}
\alias{mdl_compile}
\title{Compile ABC model}
\usage{
mdl_compile(model = modelfile)
}
\arguments{
\item{model}{Stan model file.}
}
\value{
Compiled stan model.
}
\description{
\code{mdl_compile} compiles the given stan model.
}
