% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_modelling.R
\name{mdl_core}
\alias{mdl_core}
\title{Execute ABC model.}
\usage{
mdl_core(
  data,
  chain_nr,
  init = "random",
  model,
  nchains,
  nburn,
  niter,
  nthin,
  ...
)
}
\arguments{
\item{data}{Input data frame.}

\item{chain_nr}{NUTS chain number.}

\item{init}{Initialization method.}

\item{model}{ABC Stan model.}

\item{...}{Pass through parameters.}
}
\value{
Returns a stanfit object.
}
\description{
\code{mdl_core} calls the Stan sampler to sample from the ABC model.
It should not be called directly, and instead should only be called
through the \code{\link{mdl_run}} function.
}
