% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_modeling.R
\name{mdl_init}
\alias{mdl_init}
\title{Run the initialization model.}
\usage{
mdl_init(df, ...)
}
\arguments{
\item{df}{Input data frame.}

\item{...}{Pass through parameters.}
}
\description{
Pre-process and run a faster initial run of the model.
}
