% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99_deprecated_modeling.R
\name{mdl_inits}
\alias{mdl_inits}
\title{Generate initial parameter estimates (Deprecated).}
\usage{
mdl_inits(data, chain_nr)
}
\arguments{
\item{data}{Input data list.}

\item{chain_nr}{NUTS chain number.}
}
\description{
Generate initial parameter estimates (Deprecated).
}
