% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_modelling.R
\name{mdl_joint_prep}
\alias{mdl_joint_prep}
\title{Additional ABC model inputs for joint modelling.}
\usage{
mdl_joint_prep(inputs, jpars = NULL, ...)
}
\arguments{
\item{inputs}{Pre-processed ABC model input.}

\item{jpars}{Optional joint modelling parameters.}

\item{...}{Pass through parameters.}
}
\value{
Outputs a list formatted as input into the ABC joint Stan model.
}
\description{
\code{mdl_joint_prep} appends a matrix of population proportions to the pre-processed
model input thus completing the input list for the joint model.
}
