% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99_deprecated_processing.R
\name{mdl_observed}
\alias{mdl_observed}
\title{Extract observed values (Deprecated).}
\usage{
mdl_observed(df, cutoff = 0.98)
}
\arguments{
\item{df}{Input data frame.}

\item{cutoff}{Extreme value threshold.}
}
\description{
Extract observed values (Deprecated).
}
