% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99_deprecated_plotting.R
\name{mdl_plot}
\alias{mdl_plot}
\title{Core Plotting Function for ABC Results (Deprecated)}
\usage{
mdl_plot(mdl_out, country_selection, target_yr = 2020)
}
\arguments{
\item{mdl_out}{Model output (Processed by deprecated mdl_process).}

\item{country_selection}{Vector of country codes to plot.}

\item{target_yr}{Final year to plot, defaults to 2020.}
}
\description{
Core Plotting Function for ABC Results (Deprecated)
}
