% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_plotting.R
\name{mdl_plot_gender}
\alias{mdl_plot_gender}
\title{Plot ABC Results by Gender (Deprecated)}
\usage{
mdl_plot_gender(mdl_out_g, ...)
}
\arguments{
\item{mdl_out_g}{Model output (Processed by deprecated mdl_process).}

\item{...}{Pass through parameters.}
}
\description{
Plot ABC Results by Gender (Deprecated)
}
