% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_modelling.R
\name{mdl_prep}
\alias{mdl_prep}
\title{Pre-process ABC model input.}
\usage{
mdl_prep(df, cutoff = 0.98, ...)
}
\arguments{
\item{df}{Raw input data frame produced by \code{\link{mk_dataset_4modeling}}.}

\item{cutoff}{Extreme value cap (DEPRECATED).}

\item{...}{Pass through parameters.}
}
\value{
Outputs a list formatted as input into the ABC Stan model.
}
\description{
\code{mdl_prep} is the core data pre-processing function for the ABC model
pipeline. It takes a data frame constructed by \code{\link{mk_dataset_4modeling}},
adds on additional pieces needed by the model such as indicators for late and
very late completion, and transforms the data from the original space to the
probit space. Finally, the data is converted to a list suitable to \code{\link[rstan]{stan}}
modeling using the \code{\link[tidybayes]{compose_data}} function.
}
