% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_processing.R
\name{mdl_process}
\alias{mdl_process}
\title{Process ABC model output.}
\usage{
mdl_process(df, raw_mcmc)
}
\arguments{
\item{df}{Input data frame.}

\item{raw_mcmc}{Raw ABC model MCMC output.}
}
\value{
Data frame with columns for country, year, series, value, lower,
upper, vairable, and sex. Note that series corresponds to one of:
\enumerate{
    \item projected3t5 (Official completion rate indicator)
    \item projected5 (Upper bracket of the indicator range)
    \item projected8 (Ultimate completion)
}
}
\description{
\code{mdl_process} operates on the raw stanfit object outputted by the ABC
model and summarises the findings in a digestible format. Specifically,
variants of the latent true completion rate are computed and presented in
90\% intervals.
}
