% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_processing.R
\name{mdl_process2}
\alias{mdl_process2}
\title{Process ABC model output.}
\usage{
mdl_process2(df, raw_mcmc, rounds)
}
\arguments{
\item{df}{Input data frame.}

\item{raw_mcmc}{Raw MCMC output.}

\item{rounds}{Survey rounds data frame.}
}
\description{
Process ABC model output.
}
