% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_processing.R
\name{mdl_process_dropout}
\alias{mdl_process_dropout}
\title{Process ABC model output for dropout model.}
\usage{
mdl_process_dropout(df, raw_mcmc)
}
\arguments{
\item{df}{Input data frame.}

\item{raw_mcmc}{Raw ABC model MCMC output.}
}
\value{
Data frame with columns for country, year, series, value, lower,
upper, variable, and sex. Note that series corresponds to one of:
}
\description{
\code{mdl_process_dropout} operates on the raw stanfit object outputted by the ABC
model and summarises the findings in a digestible format. Specifically,
variants of the latent true completion rate are computed and presented in
90\% intervals.
}
