% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_modelling.R
\name{mdl_run}
\alias{mdl_run}
\title{Run the ABC model.}
\usage{
mdl_run(df, ...)
}
\arguments{
\item{df}{Input data frame.}

\item{...}{Pass through parameters.}
}
\value{
Returns a stanfit object.
}
\description{
\code{mdl_run} pre-processes the input data using \code{\link{mdl_prep}}
and subsequently samples from the ABC model using \code{\link{mdl_core}}.
}
