% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_testing.R
\name{mdl_test}
\alias{mdl_test}
\title{Test model on new data}
\usage{
mdl_test(input, testset, raw_mcmc, model, type)
}
\arguments{
\item{input}{Input data frame.}

\item{testset}{Test data frame.}

\item{raw_mcmc}{Raw Stan model output.}

\item{model}{Generated quantities model.}

\item{type}{Type of test.}
}
\value{
Returns a stanfit object.
}
\description{
\code{mdl_test} uses the results of a model run to test construct the posterior
predictive distributions for new data.
}
