% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99_deprecated_modelling.R
\name{mk_agg_age_by_survey}
\alias{mk_agg_age_by_survey}
\title{Make input data by survey (Deprecated).}
\usage{
mk_agg_age_by_survey(df)
}
\arguments{
\item{df}{Raw data.}
}
\description{
Make input data by survey (Deprecated).
}
