% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7_plotting.R
\name{mk_bias_plot}
\alias{mk_bias_plot}
\title{Plot ABC model bias terms}
\usage{
mk_bias_plot(iters)
}
\arguments{
\item{iters}{Survey bias iterations.}
}
\description{
\code{mk_bias_plot} presents survey bias values at each iteration of
the model as 90% point ranges. The bias terms are grouped according
to the survey source and wave in an effort to present differences,
if any, between said sources/waves.
}
\seealso{
Other plotting functions: 
\code{\link{mk_input_plot}()},
\code{\link{mk_input_survey_plot}()},
\code{\link{mk_par_plot}()},
\code{\link{mk_reg_plot}()},
\code{\link{mk_repl_plot}()},
\code{\link{mk_rhats_plot}()},
\code{\link{mk_trace_plot}()},
\code{\link{plt_cat}()},
\code{\link{plt_country}()},
\code{\link{se_plotter}()}
}
\concept{plotting functions}
