% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_modelling.R
\name{mk_dataset_4modeling}
\alias{mk_dataset_4modeling}
\title{Construct ABC model input.}
\usage{
mk_dataset_4modeling(rounds, agedat, genderdat, ageref)
}
\arguments{
\item{rounds}{Survey wave data frame.}

\item{agedat}{Aggregate age raw data (agg_age).}

\item{genderdat}{Aggregate age by gender raw data (agg_age_gender).}

\item{ageref}{Reference ages (refages).}
}
\value{
Raw input data frame.
}
\description{
\code{mk_dataset_4modeling} is the core ABC model data compilation function.
It joins a summary of the surveys of interest with the aggregate survey data
to produce a complete data frame spanning all levels and sex combinations.
Finally, columns detailing the data reconstruction process are appended.
}
