% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_plotting.R
\name{mk_drift_plot}
\alias{mk_drift_plot}
\title{Plot long-term drift (Deprecated)}
\usage{
mk_drift_plot(df)
}
\arguments{
\item{df}{Model output (Processed by deprecated mdl_process).}
}
\description{
Plot long-term drift (Deprecated)
}
