% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7_plotting.R
\name{mk_input_plot}
\alias{mk_input_plot}
\title{Create an ABC input backprojection plot}
\usage{
mk_input_plot(input, countries)
}
\arguments{
\item{input}{Input data frame.}

\item{countries}{Vector of countries to plot.}
}
\description{
\code{mk_input_plot} plots the input data for a series of countries with
separate series for each input survey and identifiers for the age observed
and backprojection distance. It is intended to provide a complete picture
of the reconstructed surveys for a given country.
}
\seealso{
Other plotting functions: 
\code{\link{mk_bias_plot}()},
\code{\link{mk_input_survey_plot}()},
\code{\link{mk_par_plot}()},
\code{\link{mk_reg_plot}()},
\code{\link{mk_repl_plot}()},
\code{\link{mk_rhats_plot}()},
\code{\link{mk_trace_plot}()},
\code{\link{plt_cat}()},
\code{\link{plt_country}()},
\code{\link{se_plotter}()}
}
\concept{plotting functions}
