% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7_plotting.R
\name{mk_input_survey_plot}
\alias{mk_input_survey_plot}
\title{Create an ABC input survey plot}
\usage{
mk_input_survey_plot(input, survs)
}
\arguments{
\item{input}{Input data frame.}

\item{survs}{Vector of surveys to plot.}
}
\description{
\code{mk_input_survey_plot} plots the input data for a series of surveys with
identifiers for the age observed, backprojection distance, and observations
at risk of age misreporting distortions.
}
\seealso{
Other plotting functions: 
\code{\link{mk_bias_plot}()},
\code{\link{mk_input_plot}()},
\code{\link{mk_par_plot}()},
\code{\link{mk_reg_plot}()},
\code{\link{mk_repl_plot}()},
\code{\link{mk_rhats_plot}()},
\code{\link{mk_trace_plot}()},
\code{\link{plt_cat}()},
\code{\link{plt_country}()},
\code{\link{se_plotter}()}
}
\concept{plotting functions}
