% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_plotfuns.R
\name{mk_par_plot}
\alias{mk_par_plot}
\title{Plot ABC model parameters}
\usage{
mk_par_plot(df)
}
\arguments{
\item{df}{Parameter interval data frame.}
}
\description{
\code{mk_par_plot} presents boxplots of each bias/distortion parameter category
to illustrate the magnitude of each effect. The parameters in question are:
\enumerate{
  \item Survey bias
  \item Late completion
  \item Very late completion
  \item Age misreporting
}
}
\seealso{
Other plotting functions: 
\code{\link{mk_bias_plot}()},
\code{\link{mk_input_plot}()},
\code{\link{mk_input_survey_plot}()},
\code{\link{mk_reg_plot}()},
\code{\link{mk_repl_plot}()},
\code{\link{mk_rhats_plot}()},
\code{\link{mk_trace_plot}()},
\code{\link{plt_cat}()},
\code{\link{plt_country}()},
\code{\link{se_plotter}()}
}
\concept{plotting functions}
