% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7_plotting.R
\name{mk_reg_plot}
\alias{mk_reg_plot}
\title{Plot ABC Results by Region}
\usage{
mk_reg_plot(df, which_aggs = "regions", endyr = 2020)
}
\arguments{
\item{df}{Regional aggregate data frame.}

\item{which_aggs}{Type of aggregation (regions, income, or regionsXincome).}

\item{endyr}{Final year to plot.}
}
\description{
\code{mk_reg_plot} plots aggregated ABC model output using either a regional,
income level, or mixed aggregation scheme. The outputs present both the
official completion rate indicator as well as an estimate of ultimate completion.
\code{mk_reg_plot} can be used to produce plots that include forecasts reaching
to 2030 though it is worth noting that said forecasts use a mixed within-level
and across-levels weighting scheme to recognize the growth across levels is
linked in the long term.
}
\seealso{
Other plotting functions: 
\code{\link{mk_bias_plot}()},
\code{\link{mk_input_plot}()},
\code{\link{mk_input_survey_plot}()},
\code{\link{mk_par_plot}()},
\code{\link{mk_repl_plot}()},
\code{\link{mk_rhats_plot}()},
\code{\link{mk_trace_plot}()},
\code{\link{plt_cat}()},
\code{\link{plt_country}()},
\code{\link{se_plotter}()}
}
\concept{plotting functions}
