% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7_plotting.R
\name{mk_repl_plot}
\alias{mk_repl_plot}
\title{Plot Observation Replications}
\usage{
mk_repl_plot(obs, reps)
}
\arguments{
\item{obs}{Observed values data frame.}

\item{reps}{Replications data frame.}
}
\description{
\code{mk_repl_plot} creates a plot of 2 observed values from each observed
country along with the respective replications generated by the ABC model.
The observations are plotted in ascending order of value with the range of
the replications around them.
}
\seealso{
Other plotting functions: 
\code{\link{mk_bias_plot}()},
\code{\link{mk_input_plot}()},
\code{\link{mk_input_survey_plot}()},
\code{\link{mk_par_plot}()},
\code{\link{mk_reg_plot}()},
\code{\link{mk_rhats_plot}()},
\code{\link{mk_trace_plot}()},
\code{\link{plt_cat}()},
\code{\link{plt_country}()},
\code{\link{se_plotter}()}
}
\concept{plotting functions}
