% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_plotting.R
\name{mk_repl_plot2}
\alias{mk_repl_plot2}
\title{Observation replication plot (Deprecated)}
\usage{
mk_repl_plot2(tgt, df)
}
\arguments{
\item{tgt}{Observed values data frame.}

\item{df}{Replications data frame.}
}
\description{
Observation replication plot (Deprecated)
}
