% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7_plotting.R
\name{mk_rhats_plot}
\alias{mk_rhats_plot}
\title{Plot ABC model Rhats}
\usage{
mk_rhats_plot(rhats_all)
}
\arguments{
\item{rhats_all}{Model output rhats.}
}
\description{
\code{mk_rhats_plot} plots the worse case scenario Rhats for the ABC model output, grouped
by level and sex.
}
\seealso{
Other plotting functions: 
\code{\link{mk_bias_plot}()},
\code{\link{mk_input_plot}()},
\code{\link{mk_input_survey_plot}()},
\code{\link{mk_par_plot}()},
\code{\link{mk_reg_plot}()},
\code{\link{mk_repl_plot}()},
\code{\link{mk_trace_plot}()},
\code{\link{plt_cat}()},
\code{\link{plt_country}()},
\code{\link{se_plotter}()}
}
\concept{plotting functions}
