% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_plotfuns.R
\name{mk_trace_plot}
\alias{mk_trace_plot}
\title{Plot ABC model trace plots}
\usage{
mk_trace_plot(mcmc, rhats_all, lvl = "prim", sx = "total")
}
\arguments{
\item{mcmc}{ABC model mcmc output.}

\item{rhats_all}{Model output rhats.}

\item{lvl}{Level to plot.}

\item{sx}{Sex to plot.}
}
\description{
\code{mk_trace_plot} plots the trace plots of the worst case scenario parameters
as decided by the rhats diagnostic.
}
\seealso{
Other plotting functions: 
\code{\link{mk_bias_plot}()},
\code{\link{mk_input_plot}()},
\code{\link{mk_input_survey_plot}()},
\code{\link{mk_par_plot}()},
\code{\link{mk_reg_plot}()},
\code{\link{mk_repl_plot}()},
\code{\link{mk_rhats_plot}()},
\code{\link{plt_cat}()},
\code{\link{plt_country}()},
\code{\link{se_plotter}()}
}
\concept{plotting functions}
