% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/8_other.R
\name{plot_walker}
\alias{plot_walker}
\title{Plot walker?}
\usage{
plot_walker(df, pl_fun, lbl, target_yr = 2020)
}
\arguments{
\item{df}{Output data frame.}

\item{pl_fun}{Plotting function.}

\item{lbl}{File name label.}

\item{target_yr}{Final year to plot.}
}
\description{
Saves plots for each country. Replaced by report drake target?
}
