% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_plotfuns.R
\name{plt_country}
\alias{plt_country}
\title{Plot ABC Results by Country}
\usage{
plt_country(
  mdl_proj,
  mdl_obs,
  mdl_influential,
  use_influential = TRUE,
  target,
  target_yr = 2020,
  levels = c("prim", "lsec", "usec"),
  sexes = c("female", "male", "total")
)
}
\arguments{
\item{mdl_proj}{Projected values data frame.}

\item{mdl_obs}{Observed values data frame.}

\item{mdl_influential}{Influential observations data frame.}

\item{use_influential}{Boolean to indicate whether to plot influential observations.}

\item{target}{Country code.}

\item{target_yr}{Final year to plot.}

\item{levels}{Vector of education levels (prim, lsec, and usec) to plot.}

\item{sexes}{Vector of sexes (female, male, and total) to plot.}
}
\description{
\code{plt_country} creates a faceted plot displaying the observed and
estimated true completion rates for a given country. Each facet presents
a single level/sex combination. Users are given the option to highlight
points that were deemed influential in estimating some parameter in the
ABC model.
}
\seealso{
Other plotting functions: 
\code{\link{mk_bias_plot}()},
\code{\link{mk_input_plot}()},
\code{\link{mk_input_survey_plot}()},
\code{\link{mk_par_plot}()},
\code{\link{mk_reg_plot}()},
\code{\link{mk_repl_plot}()},
\code{\link{mk_rhats_plot}()},
\code{\link{mk_trace_plot}()},
\code{\link{plt_cat}()},
\code{\link{se_plotter}()}
}
\concept{plotting functions}
