% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_testing.R
\name{pred_mse}
\alias{pred_mse}
\title{Prediction summary (Deprecated).}
\usage{
pred_mse(dfs, errtgt)
}
\arguments{
\item{dfs}{List of prediction data frames.}

\item{errtgt}{Testing target.}
}
\description{
Prediction summary (Deprecated).
}
