% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_testing.R
\name{pred_mse_ace}
\alias{pred_mse_ace}
\title{Observation level prediction summary (Deprecated).}
\usage{
pred_mse_ace(dfs, errtgt)
}
\arguments{
\item{dfs}{List of prediction data frames.}

\item{errtgt}{Testing target.}
}
\description{
Observation level prediction summary (Deprecated).
}
