% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99_deprecated_testing.R
\name{pred_mse_cre}
\alias{pred_mse_cre}
\title{Survey level prediction summary (Deprecated).}
\usage{
pred_mse_cre(dfs, errtgt)
}
\arguments{
\item{dfs}{List of prediction data frames.}

\item{errtgt}{Testing target.}
}
\description{
Survey level prediction summary (Deprecated).
}
