% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explorar_estratos.R
\name{probabilidad_cae}
\alias{probabilidad_cae}
\title{Title}
\usage{
probabilidad_cae(marco_muestral, id_cae, frac_m, estratos, muestras = 1000)
}
\arguments{
\item{muestras}{}
}
\value{

}
\description{
Title
}
