% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6_aggregation.R
\name{proj2030}
\alias{proj2030}
\title{Projections to 2030}
\usage{
proj2030(df, horizon = 10)
}
\arguments{
\item{df}{Aggregated data frame.}

\item{horizon}{Projection horizon (defaults to 10).}
}
\value{
An extended data frame formatted for plotting that includes the projected values.
}
\description{
\code{proj2030} projects aggregated completion rate estimates forward to the
year 2030. Projections use a weighted average of within-level and across-level
long-term drifts with weights shifting linearly from the former to the latter
across years. Such a decision is motivated by the understanding that in the
longer term, there is likely to interactions between the levels that is not
captured in the short term by the independent modeling. For example, large
growth in primary education may eventually translate into greater growth in
the secondary levels in the long term.
}
