% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_contents.R
\name{pull_version}
\alias{pull_version}
\title{Identify the ABC version number from an ABC file}
\usage{
pull_version(.lines)
}
\arguments{
\item{.lines}{}
}
\value{

}
\description{
Identify the ABC version number from an ABC file
}
