% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6_aggregation.R
\name{reg_aggs}
\alias{reg_aggs}
\title{Regional aggregation}
\usage{
reg_aggs(df, pops)
}
\arguments{
\item{df}{Input data frame.}

\item{pops}{Population weights.}
}
\description{
Aggregates ABC data based on population weights by country and level.
}
