% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_functions.R
\name{remove_latest}
\alias{remove_latest}
\title{Remove latest survey by country.}
\usage{
remove_latest(df, type = "filter")
}
\arguments{
\item{df}{Input data frame.}

\item{type}{Filter type.}
}
\description{
Removes the latest survey from the data for countries with more than one survey. If
the user for any reason does not want to filter out any observations (such as while
static branching with drake), \code{remove_latest} can simply act as an identity
function.
}
\seealso{
Other testing functions: 
\code{\link{alt1_smpl2}()},
\code{\link{alt1_smpl3}()},
\code{\link{alt2_flat2}()},
\code{\link{alt2_flat3}()},
\code{\link{alt3_ltst3}()},
\code{\link{remove_recent}()},
\code{\link{test_abc2}()},
\code{\link{test_abc}()}
}
\concept{testing functions}
