% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_plotfuns.R
\name{se_plotter}
\alias{se_plotter}
\title{Plot standard error}
\usage{
se_plotter(df)
}
\arguments{
\item{df}{Input data frame.}
}
\description{
\code{se_plotter} plots the observed sampling standard errors on
the observed scale in a histogram.
}
\seealso{
Other plotting functions: 
\code{\link{mk_bias_plot}()},
\code{\link{mk_input_plot}()},
\code{\link{mk_input_survey_plot}()},
\code{\link{mk_par_plot}()},
\code{\link{mk_reg_plot}()},
\code{\link{mk_repl_plot}()},
\code{\link{mk_rhats_plot}()},
\code{\link{mk_trace_plot}()},
\code{\link{plt_cat}()},
\code{\link{plt_country}()}
}
\concept{plotting functions}
