% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smc.R
\name{Smc}
\alias{Smc}
\title{Performs a run using a Sequential Markov Chain (SMC) method.}
\usage{
Smc(smc.configuration, max.iterations = 40, alpha = 0.9,
  number.of.particles = 100, resample.ratio = 0.5, start.epsilon = 10,
  stop.epsilon = 0.1, number.of.replicates = 1, verbose = F)
}
\description{
This is an implementation of a SMC method. It will do a run
based on the input parameters.
}

