% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_plotting.R
\name{smpl_plot}
\alias{smpl_plot}
\title{Plot alternate model outputs (Deprecated)}
\usage{
smpl_plot(df, country_selection)
}
\arguments{
\item{df}{Simple model output data frame.}

\item{country_selection}{Vector of countries to plot.}
}
\description{
Plot alternate model outputs (Deprecated)
}
