% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_contents.R
\name{split_parts}
\alias{split_parts}
\title{Split a tune into the header and the tune body}
\usage{
split_parts(.tune)
}
\arguments{
\item{.tune}{}
}
\value{
a list with parts \verb{$header} and \verb{$body}.
}
\description{
Split a tune into the header and the tune body
}
