% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abc_examples.R
\docType{data}
\name{tennessee_wagoner}
\alias{tennessee_wagoner}
\alias{abc_tunes}
\title{Examples of simple ABC tunes}
\format{
An object of class \code{character} of length 13.
}
\usage{
tennessee_wagoner
}
\description{
Several ABC-formatted tunes are loaded as part of the package. They are
accessed by file names, stripped of the ".abc" filetype.  Accessing the tunes
this way provides a data structure consisting of a list of lines of text.
}
\details{
Available tunes include:

rights_of_man.abc

tennessee_wagoner.abc
}
\keyword{datasets}
