% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abc-view.R
\name{view_abc_file}
\alias{view_abc_file}
\title{View ABC file}
\usage{
view_abc_file(x, play_midi = FALSE, download_midi = FALSE,
  staff_width = 740)
}
\arguments{
\item{x}{Path to the ABC file.}

\item{play_midi}{(Logical scalar)
If \code{TRUE}, the display window will include a prompt whereby the
user can play an audio rendition of the musical sequence.}

\item{download_midi}{(Logical scalar)
If \code{TRUE}, the display window will provide the user with an option
to download a MIDI file representing the musical sequence.}

\item{staff_width}{(Numeric scalar)
\code{staffwidth} parameter passed to \code{ABCJS.renderAbc}.}
}
\description{
Creates an HTML preview of an ABC file.
This preview should automatically appear in the RStudio viewer pane.
}
