% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abc-view.R
\name{view_abc_string}
\alias{view_abc_string}
\title{View ABC string}
\usage{
view_abc_string(x, play_midi = FALSE, download_midi = FALSE,
  staff_width = 740)
}
\arguments{
\item{x}{(Character scalar)
ABC notation to view, expressed as a string.}

\item{play_midi}{(Logical scalar)
If \code{TRUE}, the display window will include a prompt whereby the
user can play an audio rendition of the musical sequence.}

\item{download_midi}{(Logical scalar)
If \code{TRUE}, the display window will provide the user with an option
to download a MIDI file representing the musical sequence.}

\item{staff_width}{(Numeric scalar)
\code{staffwidth} parameter passed to \code{ABCJS.renderAbc}.}
}
\description{
Creates an HTML preview of a piece of ABC notation expressed as a string.
This preview should automatically appear in the RStudio viewer pane.
}
\examples{
view_abc_string("C D E F G")
view_abc_string("L:1\\nC D E F G")
view_abc_string("L:1\\n[CEG]\\n")

}
