% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pi-chord.R
\name{view_pi_chord}
\alias{view_pi_chord}
\title{View a pitch chord}
\usage{
view_pi_chord(x, play_midi = FALSE, download_midi = FALSE,
  staff_width = 100)
}
\arguments{
\item{x}{Chord to display, expressed as a numeric vector of MIDI note numbers.
Only integerish note numbers are supported.}

\item{play_midi}{(Logical scalar)
If \code{TRUE}, the display window will include a prompt whereby the
user can play an audio rendition of the musical sequence.}

\item{download_midi}{(Logical scalar)
If \code{TRUE}, the display window will provide the user with an option
to download a MIDI file representing the musical sequence.}

\item{staff_width}{(Numeric scalar)
\code{staffwidth} parameter passed to \code{ABCJS.renderAbc}.}
}
\description{
Displays a musical chord, defined as a finite set of musical pitches.
}
