% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pi-chord-seq.R
\name{view_pi_chord_seq}
\alias{view_pi_chord_seq}
\title{View a sequence of pitch chords}
\usage{
view_pi_chord_seq(x, annotate = NULL, chords_per_line = 5,
  staff_width = 75 * pmin(pmax(2, length(x)), 5), play_midi = FALSE,
  download_midi = FALSE)
}
\arguments{
\item{x}{Chord sequence to display, expressed as a list of numeric vectors,
where each vector expresses a single chord as a list of MIDI
note numbers.}

\item{annotate}{If not \code{NULL}, a character vector of the same length as \code{x},
providing text to display under each chord. This text may not
contain whitespace.}

\item{chords_per_line}{(Integerish scalar)
Number of chords to display on each line.}

\item{staff_width}{(Numeric scalar)
\code{staffwidth} parameter passed to \code{ABCJS.renderAbc}.}

\item{play_midi}{(Logical scalar)
If \code{TRUE}, the display window will include a prompt whereby the
user can play an audio rendition of the musical sequence.}

\item{download_midi}{(Logical scalar)
If \code{TRUE}, the display window will provide the user with an option
to download a MIDI file representing the musical sequence.}
}
\description{
Displays a chord sequence, where each chord is expressed as a
vector of MIDI note numbers.
}
