% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smcToyExampleVisualisation.R
\name{VisualiseToyExampleState}
\alias{VisualiseToyExampleState}
\title{Visualise toy example state}
\usage{
VisualiseToyExampleState(smc.result, state.to.visualise = -1)
}
\description{
Visualise toy example state
}

