% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_auxiliaries.R
\name{wide_aggregate}
\alias{wide_aggregate}
\title{Aggregate Over All Category Sets}
\usage{
wide_aggregate(df, categories = "", depth = 1, estimate_se = FALSE)
}
\arguments{
\item{df}{Input data frame.}

\item{categories}{Aggregation categories.}

\item{depth}{Tree depth.}

\item{estimate_se}{Boolean indicating whether to estimate standard errors.}
}
\description{
Aggregate Over All Category Sets
}
