% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_functions.R
\name{wide_aggregate_by_cats}
\alias{wide_aggregate_by_cats}
\title{Aggregate for a given category set}
\usage{
wide_aggregate_by_cats(df, cs, estimate_se = FALSE)
}
\arguments{
\item{df}{Input data frame.}

\item{cs}{Categories.}

\item{estimate_se}{Boolean indicating whether to estimate standard errors.}
}
\description{
Aggregate for a given category set
}
