% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_modelling.R
\name{wide_bkprj_4mdl}
\alias{wide_bkprj_4mdl}
\title{Append survey and reconstruction details.}
\usage{
wide_bkprj_4mdl(df)
}
\arguments{
\item{df}{Raw input data frame.}
}
\description{
Append survey and reconstruction details.
}
