% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_functions.R
\name{wide_jk}
\alias{wide_jk}
\title{Jackknife variance estimation}
\usage{
wide_jk(d, estimate_se = FALSE)
}
\arguments{
\item{d}{Input data frame.}

\item{estimate_se}{Boolean indicating whether to estimate standard errors.}
}
\description{
Jackknife variance estimation
}
