% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ajustar_modelo.R
\name{ajustar_modelo}
\alias{ajustar_modelo}
\title{Ajusta el modelo Mendoza, Nieto-Barajas (2016)}
\usage{
ajustar_modelo(id_estratos, muestra, marco_muestral, candidatos, n_sim = 10000)
}
\arguments{
\item{n_sim}{}
}
\value{

}
\description{
Ajusta el modelo Mendoza, Nieto-Barajas (2016)
}
