% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99_deprecated_modelling.R
\name{mdl_inits}
\alias{mdl_inits}
\title{Generate initial parameter estimates (Deprecated).}
\usage{
mdl_inits(data, chain_nr, modelfile, nchains, nthin, ...)
}
\arguments{
\item{data}{Input data list.}

\item{chain_nr}{NUTS chain number.}

\item{modelfile}{Path to Stan model.}

\item{nchains}{Number of chains.}

\item{nthin}{Chain thinning parameter.}

\item{...}{Pass through parameters.}
}
\description{
Generate initial parameter estimates (Deprecated).
}
