% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_functions.R
\name{mk_dataset_4modeling}
\alias{mk_dataset_4modeling}
\title{Construct ABC model input.}
\usage{
mk_dataset_4modeling(rounds)
}
\arguments{
\item{rounds}{Survey wave data frame.}
}
\value{
Raw input data frame.
}
\description{
\code{mk_dataset_4modeling} is the core ABC model data compilation function.
It joins a summary of the surveys of interest with the aggregate survey data
to produce a complete data frame spanning all levels and sex combinations.
Finally, columns detailing the data reconstruction process are appended.
}
