% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_testing.R
\name{test_cov}
\alias{test_cov}
\title{Compute coverage probabilities}
\usage{
test_cov(input, raw_mcmc, type)
}
\arguments{
\item{input}{Input data frame.}

\item{raw_mcmc}{Raw Stan model output.}

\item{type}{Test type.}
}
\value{
Coverage probabilities.
}
\description{
Compute coverage probabilities
}
